Solve 2015-02 (#11)

+++ puzzles/2015/2015-01/src/part_one.rs
@@ -1,4 +1,4 @@
-pub fn part_one(input: &str) -> i16 {
+pub fn part_one(input: &str) -> isize {
     input
         .trim()
         .chars()
